+++ changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Extend services using the $container->extend($service, $extension, $parameter) method.
 
 ## [0.3.1] - 2019-12-31
 ### Fixed